% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelNormalize.R
\name{.SplitSamplesByMetaclusters}
\alias{.SplitSamplesByMetaclusters}
\title{Split cytometry samples by metaclusters}
\usage{
.SplitSamplesByMetaclusters(
  fnames,
  batches,
  fsom,
  fpath_out,
  cols = NULL,
  cores = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fnames}{vector. Full paths to FCS files that should be split}

\item{batches}{vector. Batch labels per file in \code{fnames}}

\item{fsom}{FlowSOM object. Trained FlowSOM model with channels matching
those in all FCS files (same panel of markers) to use for distinguishing
cell compartments (metaclusters) to normalize separately}

\item{fpath_out}{string. Path to directory where the split FCS files should
be saved}

\item{cols}{integer or string vector or \code{NULL}. Indices or names of channels
which should be normalized, or \code{NULL} to use all. Defaults to \code{NULL}}

\item{cores}{integer. Number of CPU cores to use for parallelization (at
least 2). Defaults to number of detectable cores minus 1}

\item{verbose}{logical. Whether to indicate progress. Defaults to \code{TRUE}}

\item{...}{optional additional named parameters for \code{\link[flowCore:read.FCS]{flowCore::read.FCS()}}}
}
\value{
data frame with columns \emph{'Metacluster'}, \emph{'FileName'}, \emph{'Sample'}
(\emph{i.e.}, the original file name), and \emph{'Batch'} that identifies the split FCS
files written to disk
}
\description{
Splits FCS files into metaclusters defined by a trained FlowSOM model.
}
\details{
This function uses parallelization via a SNOW cluster for speed-up.

This is an internal cytoSNOW function that does not need to be called by the
user. Input validation is not carried out, unlike in the (safer) user-level
functions \code{\link[=ParallelNormalize.Train]{ParallelNormalize.Train()}} and
\code{\link[=ParallelNormalize.apply]{ParallelNormalize.apply()}}.

The original FCS file names are used, along with the postfix
\emph{'_metaclusterXX'} where \emph{XX} stands for the metacluster number.
}
\seealso{
\code{\link[=ParallelNormalize.Train]{ParallelNormalize.Train()}} for parallel training of
CytoNorm normalization models and \code{\link[=ParallelNormalize.apply]{ParallelNormalize.apply()}} for
the parallel application of CytoNorm normalization models
}
