% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelPreprocess.R
\name{MatchChannels}
\alias{MatchChannels}
\title{Match channel names}
\usage{
MatchChannels(ref, real)
}
\arguments{
\item{ref}{string vector. Tentative channel names to compare against the
ones in the FCS files (\emph{e.g.}, from a \code{flowCore::transformList})}

\item{real}{string vector. Channel names from the FCS files (retrievable
with \code{flowCore::colnames})}
}
\value{
\code{data.frame} with columns \emph{'Reference'}, \emph{'Real'}, and \emph{'Index'}
(mapping of \code{ref} onto \code{real})
}
\description{
Uses basic string matching to align two sets of channel names that should in
principle clearly map onto each other (same panel).
For example, if a \emph{'Comp-'} prefix has been added to channel names stored in
an FCS file, the channel names specified in a \code{flowCore::transformList}
without the prefix are still clearly matchable to them.
}
\details{
Multiple matching schemes are used, to try and find a 1-to-1 matching.
The channels are matched first by their numeric part.
If this is insufficient, they are matched alphanumerically (by a letter
followed by numbers, as commonly used in naming channels).
If this is insufficient, an exact match is sought.
If all three of these approaches fail to yield a 1-to-1 matching, an error
message is generated.
}
