% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelAggregate.R
\name{ParallelAggregate}
\alias{ParallelAggregate}
\title{Advanced parallel flowFrame aggregation}
\usage{
ParallelAggregate(
  fnames,
  N,
  batches = NULL,
  cols = NULL,
  keep_file_order = FALSE,
  keep_cell_order = FALSE,
  seed = NULL,
  cores = parallel::detectCores() - 1,
  as_flowFrame = TRUE,
  guid = "Aggregate.fcs",
  fsom = NULL,
  metacluster = 1,
  verbose = TRUE
)
}
\arguments{
\item{fnames}{character vector. Full paths to FCS files.}

\item{N}{integer or Inf or NULL. Target aggregate size (>2), or Inf or NULL
to take all events. Defaults to NULL.}

\item{batches}{integer vector or NULL. Batch numbers per FCS file. Defaults
to NULL.}

\item{cols}{integer or character vector or NULL. Indices or names of channels
to use, or NULL to use all. Defaults to NULL.}

\item{keep_file_order}{logical. Whether to keep the given order of FCS files
in the aggregate. Defaults to FALSE.}

\item{keep_cell_order}{logical. Whether to keep the order of cells within
each sample in the aggregate. If N is Inf or NULL, implicitly TRUE (due to no
subsampling). Defaults to FALSE.}

\item{seed}{integer or NULL. Seed for random number generation for
subsampling, or NULL to not set a seed. Defaults to NULL.}

\item{cores}{integer. Number of CPU cores to use for multi-threading (at
least 2). Defaults to number of detectable cores minus 1.}

\item{as_flowFrame}{logical. Whether to return a flowCore::flowFrame rather
than an expression matrix with channel names for columns. Defaults to TRUE.}

\item{guid}{character. Value of GUID descriptor if flowFrame is returned.
Defaults to 'Aggregate.fcs'.}

\item{fsom}{FlowSOM object or NULL. FlowSOM model with metaclustering with
channels matching to those in all FCS files, to only extract events that map
onto a specific metacluster, or NULL to not filter events that way. Defaults
to NULL.}

\item{metacluster}{integer. Metacluster number to extract events for if
FlowSOM model is specified. Defaults to 1.}

\item{verbose}{logical. Whether to indicate progress. Defaults to TRUE.}
}
\value{
flowCore::flowFrame or matrix with named columns
}
\description{
Creates a flowFrame or matrix with named columns that aggregates data from
specified FCS files.
The size of this aggregate may be capped at an approximately fixed size.
The process uses parallelisation via a SNOW cluster for speed-up.
}
\details{
For a target aggregate size N and n FCS files, each flowFrame contributes
about N/n, unless one or more files have fewer than N/n cells, in which case
all their events are taken and the sample sizes for other files is increased
to compensate for this.

If a corresponding FlowSOM model with metaclustering is given, we can choose
to extract only events that get mapped onto a specific metacluster.
In this case, up to N/n events per sample are picked greedily.

The aggregate gets extra columns: 'File' with index of FCS file of origin per
event and 'OriginalIndex' with index of each event within its FCS file of
origin.

If batch per file is specified, the 'Batch' column is added with the batch
number of origin per event.

The output gets attributes 'fnames', 'seed' and 'metacluster' (which gets
NULL if no FlowSOM model was given) with the corresponding argument values.
}
