% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelGetFeatures.R
\name{ParallelGetFeatures}
\alias{ParallelGetFeatures}
\title{Parallel FlowSOM feature extraction}
\usage{
ParallelGetFeatures(
  fsom,
  fnames,
  level = c("clusters", "metaclusters"),
  type = c("counts", "proportions", "medians", "phenopositivities"),
  state_markers = NULL,
  thresholds = NULL,
  cores = parallel::detectCores() - 1,
  sample_names = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fsom}{FlowSOM object. Trained FlowSOM model with channels matching to
those in all FCS files (same panel of markers)}

\item{fnames}{vector. Full paths to FCS files}

\item{level}{vector. Resolution level(s) of interest for feature
extraction. One or both of \code{c('clusters', 'metaclusters')}. Defaults to both}

\item{type}{vector. Type(s) of features to extract. One or more of
\code{c('counts', 'proportions', 'medians', 'phenopositivities')}
State feature types (medians and phenopositivities) require \code{state_markers}
to be specified. Additionally, phenopositivities require \code{thresholds} to be
specified}

\item{state_markers}{vector or NULL. Channels or markers to
consider for the extraction of state features. Required for computing medians
and/or phenopositivities. Defaults to NULL}

\item{thresholds}{named numeric vector or NULL. Expression values per channel
or marker that are the upper bounds for strictly negative phenotype. Required
for computing phenopositivities. Defaults to NULL}

\item{cores}{integer. Number of CPU cores to use for multi-threading (at
least 2). Defaults to number of detectable cores minus 1}

\item{sample_names}{string vector or NULL. Names of samples in \code{fnames} to
use as the rownames of features matrices. Otherwise, \code{fnames} themselves are
used. Defaults to NULL}

\item{verbose}{logical. Whether to indicate progress. Defaults to TRUE}

\item{...}{optional additional named parameters for \code{flowCore::read.FCS}}
}
\value{
list with named elements corresponding to matrices per requested
feature \code{type}, each with prefix \emph{'clusters_'} or \emph{'metaclusters_'}
}
\description{
Computes matrices of abundance and state features per FCS sample.
This is done by mapping the expression matrices of selected FCS samples onto
an existing FlowSOM object.
}
\details{
This function uses parallelisation via a SNOW cluster for speed-up.

Two groups of features can be extracted: \strong{abundance} features and \strong{state}
features.
\strong{Abundance} features can be used to describe compositional changes between
samples in terms of cell type representation.
Those can be tested using \strong{differential abundance} modelling in \emph{diffcyt}.
\strong{State} features can be used to describe changes changes in cell state
between samples across different cell types.
Those can be tested using \strong{differential state} modelling in \emph{diffcyt} and
\emph{iidx}.

Possible \strong{abundance} feature types are \emph{'counts'} and \emph{'proportions'}.
They correspond, respectively, to the absolute count and relative proportion
of cells per (meta)cluster in each sample.

Possible \strong{state} feature types are \emph{'medians'} and \emph{'phenopositivities'}.
Here, 'medians' correspond to the median expression values of a defined
subset of markers per (meta)cluster in each sample.
In contrast, \emph{'phenopositivities'} correspond to the proportions of cells
that exceed a threshold for phenotypically positive for a defined subset of
markers, per (meta)cluster in each sample.
The phenopositivity thresholds then need to specified manually.

The feature type names, in the order described here, correspond to
\emph{'counts'}, \emph{'percentages'}, \emph{'MFIs'}, and \emph{'percentages_positive'} in the
FlowSOM package.
}
