% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelNormalize.R
\name{ParallelNormalize.Apply}
\alias{ParallelNormalize.Apply}
\title{Parallel normalization model application}
\usage{
ParallelNormalize.Apply(
  model,
  fnames,
  batches,
  fpath_out,
  keep_full_paths = FALSE,
  prefix = "Norm_",
  normMethod.normalize = CytoNorm::QuantileNorm.normalize,
  cores = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{model}{normalization model generated with
\code{\link[=ParallelNormalize.train]{ParallelNormalize.train()}}}

\item{fnames}{vector. Full paths to FCS files which should be normalized}

\item{batches}{vector. Batch labels per file in \code{fnames}}

\item{fpath_out}{string. Path to directory to save normalized files}

\item{keep_full_paths}{logical. Whether to include the full path to each FCS
file in the name of its normalized counterpart. Defaults to FALSE}

\item{prefix}{string. Prefix to glue to the beginning of the original file
names in \code{fnames} to denote normalized. Defaults to \emph{'Norm_'}}

\item{normMethod.normalize}{application function of the signal normalization
method to use. Defaults to \code{CytoNorm::QuantileNorm.normalize}}

\item{cores}{integer. Number of CPU cores to use for parallelization (at
least 2). Defaults to number of detectable cores minus 1}

\item{verbose}{logical. Whether to indicate progress. Defaults to \code{TRUE}}

\item{...}{optional additional named parameters for \code{\link[flowCore:read.FCS]{flowCore::read.FCS()}}}
}
\value{
nothing is returned
}
\description{
Applies a CytoNorm batch effect correction model, trained via
\code{\link[=ParallelNormalize.Train]{ParallelNormalize.Train()}}, to FCS files affected by batch effect.
}
\details{
This function uses parallelization via a SNOW cluster for speed-up.
}
\note{
It is highly advisable to have unique file names for each input FCS sample
and have all of them in the same directory (no nested directories). An
alternative solution is to take the entire paths to each sample as unique
identifiers (set \code{keep_full_paths} to TRUE), but this is discouraged.
}
\seealso{
\code{\link[CytoNorm:CytoNorm.normalize]{CytoNorm::CytoNorm.normalize()}} for the original CytoNorm
normalization function
}
