% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelNormalize.R
\name{ParallelNormalize.Train}
\alias{ParallelNormalize.Train}
\title{Parallel training of normalization model}
\usage{
ParallelNormalize.Train(
  fnames,
  batches,
  cols = NULL,
  fsom = NULL,
  fsom_params = list(nCells = 1e+06, xdim = 15, ydim = 15, nClus = 30, scale = FALSE),
  gating_fun = NULL,
  gate_names = NULL,
  normMethod.train = CytoNorm::QuantileNorm.train,
  norm_params = list(nQ = 99),
  seed = NULL,
  cores = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fnames}{vector. Full paths to training FCS files}

\item{batches}{vector. Batch labels per file in \code{fnames}}

\item{cols}{integer or string vector or \code{NULL}. Indices or names of channels
which should be normalized, or \code{NULL} to use all. Defaults to \code{NULL}}

\item{fsom}{FlowSOM object. Trained FlowSOM model with channels matching
those in all FCS files (same panel of markers) to use for distinguishing
cell compartments (metaclusters) to normalize separately. Or \code{NULL} to train
a new one using \code{fsom_params} as the training parameters, or to use a gating
model instead of FlowSOM. Defaults to \code{NULL}}

\item{fsom_params}{named list. Parameters to pass to
\code{\link[CytoNorm:prepareFlowSOM]{CytoNorm::prepareFlowSOM()}} for the FlowSOM model training. Ignored if
\code{fsom} is specified or using a gating model instead. If specified,
\code{fsom_params} must give at least the \emph{'nCells'} specified. See \emph{Details}
for defaults}

\item{gating_fun}{function. Takes a \link[flowCore:flowFrame-class]{flowCore::flowFrame} object as input and
returns a gate label per event. Or \code{NULL} to use FlowSOM for splitting files
instead. Defaults to \code{NULL}}

\item{gate_names}{vector. Names of all gates from \code{gating_fun}. Or \code{NULL} to
use FlowSOM for splitting files. Must be specified if \code{gating_fun} is
specified. Defaults to \code{NULL}}

\item{normMethod.train}{training function of the signal normalization method
to use. Defaults to \code{CytoNorm::QuantileNorm.train}}

\item{norm_params}{named list. Parameters to pass to the normalization method
function specified in \code{norm_method}. Defaults to \code{list('nQ' = 99)}}

\item{seed}{integer or \code{NULL}. Seed for reproducibility}

\item{cores}{integer. Number of CPU cores to use for parallelization (at
least 2). Defaults to number of detectable cores minus 1}

\item{verbose}{logical. Whether to indicate progress. Defaults to \code{TRUE}}

\item{...}{optional additional named parameters for \code{\link[flowCore:read.FCS]{flowCore::read.FCS()}}}
}
\value{
list with named elements for the CytoNorm normalization model
(\emph{'clusterRes'}) and the splitting model: either FlowSOM (\emph{'fsom'}) or the
custom gating model (\emph{'gating'}).
}
\description{
Trains a batch-effect correction model using CytoNorm. This model can later
be applied using \code{\link[=ParallelNormalize.Apply]{ParallelNormalize.Apply()}}. Instead of using
FlowSOM clustering for splitting files into compartments, a custom gating
function can be used (see the \code{gating_fun} and \code{gate_names} parameters and
the \emph{Details} section).
}
\details{
This function uses parallelization via a SNOW cluster for speed-up.

If no precomputed FlowSOM model is given (\code{fsom} is not specified), the
default training parameters are \code{nCells = 1e6}, \code{xdim = 15}, \code{ydim = 15},
\code{nClus = 30}, and \code{scale = FALSE}. If \code{colsToUse} (channels to use for
clustering) are not specified here, the value of \code{cols} (all channels that
should be normalized) is used. If \code{cols} is also not specified, all channels
in the panel are used. It is strongly recommended to specify both the \code{cols}
parameter, and the \code{colsToUse} slot of \code{fsom_params}.

A gating model can be used to split FCS files instead of a FlowSOM model. To
do this, two input parameters need to be specified. First, a gating function
that takes a \link[flowCore:flowFrame-class]{flowCore::flowFrame} as input and returns a vector of gate
labels as output (\code{gating_fun}). Second, a vector with the names of all gates
that \code{gating_fun} can output (\code{gate_names}). If both are specified, the
FlowSOM model and/or parameters are ignored. Instead of splitting training
FCS files by metacluster to train a separate normalization function for each
of them, they are split by gate, as defined by the user's function. This
gives more control and ensures that the model doesn't skew genuine
differential abundances of known cell types across batches.
}
\note{
The original \code{\link[CytoNorm:CytoNorm.train]{CytoNorm::CytoNorm.train()}} function includes a \code{transformList}
parameter that lets the user apply transformation in conjunction with the
normalization model training/application. In contrast, this function does not
include that parameter. This is to encourage doing preprocessing separately
first, checking whether it's working correctly, and only then proceeding with
normalization.

Note that this function generates temporary files in the \code{\link[=tempdir]{tempdir()}}
directory, which it then automatically removes when done.

If a FlowSOM model (not gating) is used for the splitting of files, the
trained normalization model is compatible with the original
\code{\link[CytoNorm:CytoNorm.normalize]{CytoNorm::CytoNorm.normalize()}} function for applying the normalization.
However, we recommend the use of \code{\link[=ParallelNormalize.Apply]{ParallelNormalize.Apply()}}.
}
\seealso{
\code{\link[CytoNorm:CytoNorm.train]{CytoNorm::CytoNorm.train()}} for the original CytoNorm training
function

\code{\link[=ParallelNormalize.Apply]{ParallelNormalize.Apply()}} for applying the trained
normalization model
}
