% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateFCS.R
\name{ValidateFCS}
\alias{ValidateFCS}
\title{Make a valid flowFrame}
\usage{
ValidateFCS(exprs, ff = NULL, desc = NULL)
}
\arguments{
\item{exprs}{numeric matrix with named columns. Row-wise cytometry
expression matrix}

\item{ff}{\code{flowCore::flowFrame} or \code{NULL}. Template flowFrame, on which the
new flowFrame should be based. Defaults to \code{NULL}}

\item{desc}{named string vector or \code{NULL}. Template \code{desc} vector (\emph{e.g.},
\code{ff@parameters@data$desc}) to use. Defaults to \code{NULL}}
}
\value{
\code{flowCore::flowFrame}
}
\description{
Generates a \code{flowCore::flowFrame} based on an expression matrix and optional
template flowFrame and/or \code{desc} element.
Required slots of the flowFrame XML are filled in, which should prevent error
messages that can otherwise occur when the \code{exprs} slot of an existing
flowFrame gets rewritten.
}
