% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelAggregate.R
\name{ParallelAggregate}
\alias{ParallelAggregate}
\title{Parallel flowFrame aggregation}
\usage{
ParallelAggregate(
  fnames,
  N,
  batches = NULL,
  cols = NULL,
  keep_file_order = FALSE,
  keep_cell_order = FALSE,
  seed = NULL,
  cores = parallel::detectCores() - 1,
  as_flowFrame = TRUE,
  guid = "Aggregate.fcs",
  fsom = NULL,
  metacluster = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fnames}{string vector. Full paths to FCS files}

\item{N}{integer or \code{Inf} or \code{NULL}. Target aggregate size (>2), or \code{Inf} or
\code{NULL} to take all events. Defaults to \code{NULL}}

\item{batches}{integer vector or \code{NULL}. Batch numbers per FCS file. Defaults
to \code{NULL}}

\item{cols}{integer or string vector or \code{NULL}. Indices or names of
channels to use, or \code{NULL} to use all. Defaults to \code{NULL}}

\item{keep_file_order}{logical. Whether to keep the given order of \code{FCS}
files in the aggregate. Defaults to \code{FALSE}}

\item{keep_cell_order}{logical. Whether to keep the order of cells within
each sample in the aggregate. If \code{N} is \code{Inf} or \code{NULL}, implicitly \code{TRUE}
(due to no subsampling). Defaults to \code{FALSE}}

\item{seed}{integer or \code{NULL}. Seed for random number generation for
subsampling, or \code{NULL} to not set a seed. Defaults to \code{NULL}}

\item{cores}{integer. Number of CPU cores to use for multi-threading (at
least 2). Defaults to number of detectable cores minus 1}

\item{as_flowFrame}{logical. Whether to return a \code{flowCore::flowFrame} rather
than an expression matrix with channel names for columns. Defaults to \code{TRUE}}

\item{guid}{string. Value of the \code{GUID} descriptor if
\code{flowCore::flowFrame} is to be returned. Defaults to \emph{'Aggregate.fcs'}}

\item{fsom}{FlowSOM object or \code{NULL}. FlowSOM model with metaclustering with
channels matching to those in all FCS files, to only extract events that map
onto a specific metacluster, or \code{NULL} to not filter events that way.
Defaults to \code{NULL}}

\item{metacluster}{integer. Metacluster number to extract events, if FlowSOM
model is specified. Defaults to 1}

\item{verbose}{logical. Whether to indicate progress. Defaults to \code{TRUE}}

\item{...}{optional additional named parameters for \code{flowCore::read.FCS}}
}
\value{
\code{flowCore::flowFrame} or matrix with named columns
}
\description{
Creates a \code{flowCore::flowFrame} or a matrix with named columns that
aggregates data from specified FCS files.
The size of this aggregate may be capped at an approximately fixed size.
}
\details{
This function uses parallelisation via a SNOW cluster for speed-up.

For a target aggregate size \code{N} and \code{n} FCS files, each \code{flowFrame}
contributes about \code{N/n}, unless one or more files have fewer than \code{N/n}
cells, in which case all their events are taken and the sample sizes for
other files is increased to compensate for this.

If a corresponding FlowSOM model with metaclustering is given, we can choose
to extract only events that get mapped onto a specific metacluster.
In this case, up to \code{N/n} events per sample are picked greedily.

The aggregate gets extra columns: \emph{'File'} with index of FCS file of origin
per event and \emph{'OriginalIndex'} with index of each event within its FCS file
of origin.

If batch per file is specified, the \emph{'Batch'} column is added with the batch
number of origin per event.

The output gets attributes \code{fnames}, \code{seed}, and \code{metacluster} (which gets
\code{NULL} if no FlowSOM model was given) with the corresponding argument values.
These can be accessed as \code{attributes(output_object)$fnames}, etc.
}
