% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelNormalize.R
\name{ParallelNormalize.Train}
\alias{ParallelNormalize.Train}
\title{Parallel normalization model training}
\usage{
ParallelNormalize.Train(
  fnames,
  batches,
  cols = NULL,
  fsom = NULL,
  fsom_params = list(nCells = 1e+06, xdim = 15, ydim = 15, nClus = 30, scale = FALSE),
  normMethod.train = CytoNorm::QuantileNorm.train,
  norm_params = list(nQ = 99),
  seed = NULL,
  cores = parallel::detectCores() - 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fnames}{vector. Full paths to training FCS files}

\item{batches}{vector. Batch labels per file in \code{fnames}}

\item{cols}{integer or string vector or \code{NULL}. Indices or names of channels
which should be normalized, or \code{NULL} to use all. Defaults to \code{NULL}}

\item{fsom}{FlowSOM object. Trained FlowSOM model with channels matching
those in all FCS files (same panel of markers) to use for distinguishing
cell compartments (metaclusters) to normalize separately. Or \code{NULL} to train
a new one, using \code{fsom_params} as the training parameters. Defaults to \code{NULL}}

\item{fsom_params}{named list. Parameters to pass to
\code{\link[CytoNorm:prepareFlowSOM]{CytoNorm::prepareFlowSOM()}} for the FlowSOM model training. Ignored if
\code{fsom} is specified. Must have at least \emph{'nCells'} specified. See \emph{Details}
for defaults}

\item{normMethod.train}{training function of the signal normalization method
to use. Defaults to \code{CytoNorm::QuantileNorm.train}}

\item{norm_params}{named list. Parameters to pass to the normalization method
function specified in \code{norm_method}. Defaults to \code{list('nQ' = 99)}}

\item{seed}{integer or \code{NULL}. Seed for reproducibility}

\item{cores}{integer. Number of CPU cores to use for parallelization (at
least 2). Defaults to number of detectable cores minus 1}

\item{verbose}{logical. Whether to indicate progress. Defaults to \code{TRUE}}

\item{...}{optional additional named parameters for \code{\link[flowCore:read.FCS]{flowCore::read.FCS()}}}
}
\value{
list with named elements for the normalization FlowSOM model
(\emph{'fsom'}) and the CytoNorm normalization model itself (\emph{'clusterRes'}).
These names were chosen for compatibility with CytoNorm
}
\description{
Trains a batch effect correction model using CytoNorm. This model can later
be applied using \code{\link[=ParallelNormalize.Apply]{ParallelNormalize.Apply()}}.
}
\details{
This function uses parallelization via a SNOW cluster for speed-up.

If no precomputed FlowSOM model is given (\code{fsom} is not specified), the
default training parameters are \code{nCells = 1e6}, \code{xdim = 15}, \code{ydim = 15},
\code{nClus = 30}, and \code{scale = FALSE}. If \code{colsToUse} (channels to use for
clustering) are not specified here, the value of \code{cols} (all channels that
should be normalized) is used. If \code{cols} is also not specified, all channels
in the panel are used. It is strongly recommended to specify both the \code{cols}
parameter, and the \code{colsToUse} slot of \code{fsom_params}.
}
\note{
The original \code{\link[CytoNorm:CytoNorm.train]{CytoNorm::CytoNorm.train()}} function includes a \code{transformList}
parameter that lets the user apply transformation in conjunction with the
normalization model training/application. In contrast, this function does not
include that parameter. This is to encourage doing preprocessing separately
first, checking whether it's working correctly, and only then proceeding with
normalization.

Note that this function generates temporary files in the \code{\link[=tempdir]{tempdir()}}
directory, which it then automatically removes when done.
}
\seealso{
\code{\link[CytoNorm:CytoNorm.train]{CytoNorm::CytoNorm.train()}} for the original CytoNorm training
function
}
